+++ lib/FeatureFlag.php
@@ -134,9 +134,9 @@ class FeatureFlag
 
         if ($AhasVariantOverride && $BhasVariantOverride) {
             return 0;
-        } else if ($AhasVariantOverride) {
+        } elseif ($AhasVariantOverride) {
             return -1;
-        } else if ($BhasVariantOverride) {
+        } elseif ($BhasVariantOverride) {
             return 1;
         } else {
             return 0;
@@ -155,20 +155,20 @@ class FeatureFlag
             $flagConditionsWithIndexes[] = array($value, $i);
             $i++;
         }
-        // # Stable sort conditions with variant overrides to the top. This ensures that if overrides are present, they are
+        // # Stable sort conditions with variant overrides to the top.
+        // # This ensures that if overrides are present, they are
         // # evaluated first, and the variant override is applied to the first matching condition.
         usort(
             $flagConditionsWithIndexes,
-            function ($conditionA, $conditionB)
-            {
+            function ($conditionA, $conditionB) {
                 $AhasVariantOverride = isset($conditionA[0]["variant"]);
                 $BhasVariantOverride = isset($conditionB[0]["variant"]);
-        
+
                 if ($AhasVariantOverride && $BhasVariantOverride) {
                     return $conditionA[1] - $conditionB[1];
-                } else if ($AhasVariantOverride) {
+                } elseif ($AhasVariantOverride) {
                     return -1;
-                } else if ($BhasVariantOverride) {
+                } elseif ($BhasVariantOverride) {
                     return 1;
                 } else {
                     return $conditionA[1] - $conditionB[1];
@@ -206,7 +206,7 @@ class FeatureFlag
 
     private static function isConditionMatch($featureFlag, $distinctId, $condition, $properties)
     {
-        $rolloutPercentage = $condition["rollout_percentage"];
+        $rolloutPercentage = array_key_exists("rollout_percentage", $condition) ? $condition["rollout_percentage"] : null;
 
         if (count($condition['properties'] ?? []) > 0) {
             foreach ($condition['properties'] as $property) {