+++ test/FeatureFlagTest.php
@@ -729,6 +729,26 @@ class FeatureFlagMatch extends TestCase
         ]);
     }
 
+    public function testComputingFlagWithoutRolloutLocally()
+    {
+        $this->http_client = new MockedHttpClient(host: "app.posthog.com", flagEndpointResponse: MockedResponses::LOCAL_EVALUATION_WITH_NO_ROLLOUT_REQUEST);
+        $this->client = new Client(
+            FAKE_API_KEY,
+            [
+                "debug" => true,
+            ],
+            $this->http_client,
+            "test"
+        );
+        PostHog::init(null, null, $this->client);
+
+        $flags = PostHog::getAllFlags('distinct-id');
+
+        $this->assertEquals($flags, [
+            "enabled-flag" => true,
+        ]);
+    }
+
     public function testFlagWithVariantOverrides()
     {
         $this->http_client = new MockedHttpClient(host: "app.posthog.com", flagEndpointResponse: MockedResponses::LOCAL_EVALUATION_VARIANT_OVERRIDES_REQUEST);